extract out score header file
